+++ compiler/semantic-analysis/src/type_checkers/resolution.rs
@@ -33,6 +33,8 @@ impl TypeChecker {
                         Some(Type::Struct(name.to_string()))
                     } else if self.enum_defs.contains_key(name) {
                         Some(Type::Enum(name.to_string()))
+                    } else if self.newtype_defs.contains_key(name) {
+                        Some(Type::Newtype(name.to_string()))
                     } else {
                         self.record_error(TypeError::UnknownTypeName {
                             name: name.to_string(),